Remove element from state by prop
Use <filter> method to the state as array. (objects are elements of the array)
The argument of <filter> is an arrow function.

First argument of the arrow function is the identifier
which will be named to mean each other objects in the state.

Second argument of the arrow function is the identifier
which will be named to mean each other indexes of objects in the state.

If you add the other condition using first argument,
<filter> will filter the state with two conditions.

First argument works as an object identifier,
then member access to prop is available and will be removed by <filter>