Define state and callback function in context file
State and callback function for context data and the object of them are needed to be defined in context file.

Callback function can be initialized as an arrow function in initialization of <valueToShare>.

They're defined but not in a component, but in the context file.
They need to be defined in a component. What component can do it?