Define Provider component in context file
Define state, callback function, and the object to share in <Provider> component.
Specify <children> object as a parameter of <Provider> component.

Return <BookContext.Provider> component with <value> prop passed the shared object.
Wrap <children> object with the context built-in Provider.
Export <Provider> component as named export.
CAUTION : Identifiers of state and callback function must be same in sub-components to use the context system.