Initialize <deleteBookById> in <BookShow>
Initialize <deleteBookById> with object destructuring.
Remove props from the parameters of <BookShow> except <book>.

Call of <onEdit> in this component is not needed anymore.
Because <editBookById> will be called in <BookEdit> component.
So, <handleSubmit> doesn't need parameters anymore.

Replace call of <onDelete> with <deleteBookById> and input <book.id>.

<BookEdit> component should take <book> prop and <onEditSubmit>.
Because it is simpler to display the edited book.