Create <Accordion> component
Remove all code in <App>, they are copied in <ButtonPage>.
Create sample data which has an array of objects.

Objects contains <label> or <Heading> and <content> members.
Specify <items> prop to call of <Accordion> component.
Pass reference to <items> array as value of <items> prop.
Specify parameter of <Accordion> to reference to <items>.