Get <option> user clicked
Specify shorthand event handler to <onClick> prop at the return of <renderedOptions>.
Sepcify the parameter of <handleOptionClick> with <option> which will be passed through the shorthand event handler.
the mapped data of <options> will be passed to <handleOptionClick>, then you can know the value of what user-clicked in the dropdown menu.