Use <current> member of <useRef> returned
If <Dropdown> component doesn't contain event, then <isOpen> is updated to false.
So the component will be closed.

It makes user-clicked component can be closed if user clicked another component.

Sometimes, a reference to element can be NULL, then access to kind of <devEl> makes an error.
So the condition for <current> member is not null is needed.