Customize <ModalPage> and set <Modal>
Initialize <modal> variable to use in <Modal> component.
Initialize <actionBar> variable to use in <Modal> component.
Specify <actionBar> and <children> prop to <Modal> in <ModalPage>, <children> is the paragraph element.
Specify <onClose> prop to <Modal> to close the modal window if user clicked the <Button>.

Specify <children> and <actionBar> as the parameters of <Modal> component.
Specify <children> and <actionBar> as the returned JSX element in <Modal> component.